Proyecto Final

+++ Javascript Calderón/scripts/carrito.js
@@ -1,5 +1,4 @@
-let carrito2 = JSON.parse(localStorage.getItem("carrito")) || []
-localStorage.setItem("carrito2", JSON.stringify(carrito2));
+let carrito = JSON.parse(localStorage.getItem("carrito")) || []
 
 const detalle = document.querySelector("#detalle")
 const divCarrito = document.querySelector("#carrito")
@@ -8,44 +7,52 @@ const finCompra = document.querySelector("#finCompra")
 
 function productosCarrito() {
    
-    carrito2.map(producto => {
+    carrito.forEach(producto => {
         divCarrito.innerHTML += `
         <li class="detalleProd">
         <img src=".${producto.img}"/>
         <p>${producto.nombre}</p>
-        <p>${producto.precio}</p>
+        <b><p>$${producto.precio}</p></b>
         <button class="botonBorrar">X</button>
         </li>`
-
-        let total = carrito2.reduce((acc, curr) => acc + curr.precio, 0)
-    let totalCompra = document.createElement("p")
-    totalCompra.setAttribute("class", "total")
-    totalCompra.innerText = ("Total: " + total)
-    divCarrito.append(totalCompra)
 
     let botonBorrar = document.getElementsByClassName("botonBorrar")
 
     for (botonX of botonBorrar) {
         botonX.addEventListener("click", eliminarProducto)
     }
+    
 
     // vaciar carrito
     vaciar.addEventListener("click", () => {
-        carrito2 = []
-        localStorage.clear()
-        divCarrito.innerHTML = ""
-        contador = false
+        swal({
+            title: "Deseas vaciar tu carrito?",
+            text: "Si lo vacias deberas ir a la tienda nuevamente!",
+            icon: "warning",
+            buttons: true,
+            dangerMode: true,
+          })
+          .then((willDelete) => {
+            if (willDelete) {
+                carrito = []
+                localStorage.clear()
+                divCarrito.innerHTML = ""
+                contador = false
+            }
+          });
+        
     })
     })
-
+    
+    let total = carrito.reduce((acc, curr) => acc + curr.precio, 0)
+    let totalCompra = document.createElement("p")
+    totalCompra.setAttribute("class", "total")
+    totalCompra.innerText = ("Total: $" + total)
+    divCarrito.append(totalCompra)
 
 }
 
-
-productosCarrito(carrito)
-
-// el carrito tiene productos ? si tiene, mostrarlos : si no tiene mostrar mensaje
-carrito2.length ? productosCarrito() : (detalle.innerHTML=`<h3>El carrito está vacío</h3>`)
+carrito.length ? productosCarrito() : (detalle.innerHTML=`<h3>El carrito está vacío</h3>`)
 
 
 //FORMULARIO
@@ -53,20 +60,40 @@ let nombre = document.querySelector("#nombre")
 let mail = document.querySelector("#mail")
 let telefono = document.querySelector("#telefono")
 let direccion = document.querySelector("#direccion")
-let divConfirmacion = document.querySelector("#confirmacion")
+let divfinForm= document.querySelector("#finForm")
 let submit = document.querySelector("#submit")
 let formulario = document.querySelector("#form")
 
 function terminarCompra() {
 
-    if (nombre.value !== "" && mail.value !== "" && telefono.value !== "" && direccion.value !== "") {
-        divConfirmacion.innerHTML = `
-<h3>Gracias por tu compra ${nombre.value}!</h3>
-<p>En 2/3 días hábiles lo estaremos enviando a ${direccion.value}</p>
-`
-    } else {
-        divConfirmacion.innerHTML = `
-        <h3>Por favor complete todos los campos</h3>`
+    if(carrito.length == 0){
+        swal({
+            title: "ERROR!",
+            text: "No se puede comprar con el carrito vacio!",
+            icon: "error",
+            button: "Volver"
+        })
+    }
+    else if (nombre.value == "" || mail.value == "" || telefono.value == "" || direccion.value == "") {
+        swal({
+            title: "ERROR!",
+            text: "Por favor complete todos los campos!",
+            icon: "error",
+            button: "Volver"
+        })
+    }
+     else {
+        swal({
+            title: "Gracias por tu compra " + nombre.value + "!",
+            text: "En 3 días hábiles estaremos enviando tu pedido a " + direccion.value,
+            icon: "success",
+            button: "Continuar"
+        })
+          carrito = []
+          localStorage.clear()
+          divCarrito.innerHTML = ""
+          contador = false    
+        
     }
 }
 
@@ -76,17 +103,34 @@ formulario.onsubmit = (e) => {
     formulario.reset()
 }
 
-carrito.length && productosCarrito()
-
 
 //ELIMINAR PRODUCTO
 function eliminarProducto(e) {
-    divCarrito.innerHTML = ""
-    const botonX = e.target;
-    const idBotonX = botonX.getAttribute("id");
-    let indexProducto = carrito2.findIndex(producto => producto.id === idBotonX)
-    carrito2.splice(indexProducto, 1)
-    localStorage.removeItem("carrito2")
-    localStorage.setItem("carrito2", JSON.stringify(carrito2))
-    productosCarrito(carrito)
+    swal({
+        title: "Estas seguro?",
+        text: "Estas a punto de eliminar un articulo del carrito!",
+        icon: "warning",
+        buttons: true,
+        dangerMode: true,
+      })
+      .then((willDelete) => {
+        if (willDelete) {
+            divCarrito.innerHTML = ""
+            const botonX = e.target;
+            const idBotonX = botonX.getAttribute("id");
+            let indexProducto = carrito.findIndex(producto => producto.id === idBotonX)
+            carrito.splice(indexProducto, 1)
+            localStorage.removeItem("carrito")
+            localStorage.setItem("carrito", JSON.stringify(carrito))
+            productosCarrito(carrito)
+            swal("El articulo se elimino del carrito!", {
+            icon: "success",
+          });
+
+          Toastify({
+            text: "El carrito contiene " + carrito.length + " productos!",
+            close: true
+            }).showToast()
+        }
+      }); 
 }
